+++ Editor/VisionOSBuildProcessor.PreProcessor.cs
@@ -1,12 +1,15 @@
+#if UNITY_VISIONOS || (UNITY_VISIONOS_MAC_STUB && UNITY_STANDALONE_OSX)
 using System;
 using UnityEditor.Build;
 using UnityEditor.Build.Reporting;
 using UnityEngine;
+#endif
 
 namespace UnityEditor.XR.VisionOS
 {
     static partial class VisionOSBuildProcessor
     {
+#if UNITY_VISIONOS || (UNITY_VISIONOS_MAC_STUB && UNITY_STANDALONE_OSX)
         class Preprocessor : IPreprocessBuildWithReport
         {
             const string k_PreCompiledLibraryName = "libUnityVisionOS.a";
@@ -23,6 +26,7 @@ namespace UnityEditor.XR.VisionOS
 
             void IPreprocessBuildWithReport.OnPreprocessBuild(BuildReport report)
             {
+                DisableSplashScreenIfEnabled();
                 SetRuntimePluginCopyDelegate();
 
                 if (!IsLoaderEnabled())
@@ -39,6 +43,18 @@ namespace UnityEditor.XR.VisionOS
                     Debug.Log("Notice: an Unbounded volume configuration is required for ARKit features when building for Mixed Reality");
 #endif
                 }
+            }
+
+            static void DisableSplashScreenIfEnabled()
+            {
+                s_SplashScreenWasEnabled = PlayerSettings.SplashScreen.show;
+                if (!s_SplashScreenWasEnabled)
+                    return;
+
+                Debug.LogWarning("The Unity splash screen is not supported on visionOS. It will be disabled for this build and re-enabled afterward. " +
+                    "You may need to manually re-enable the splash screen in Player Settings if this build fails.");
+
+                PlayerSettings.SplashScreen.show = false;
             }
 
             static void SetRuntimePluginCopyDelegate()
@@ -78,10 +94,6 @@ namespace UnityEditor.XR.VisionOS
 
             static bool ShouldIncludePreCompiledLibraryInBuild(string path)
             {
-#if UNITY_2022_3_9 || UNITY_2022_3_10
-                // Exclude simulator library, which is not used in 2022.3.9f1 and 2022.3.10f1
-                return !path.Contains("Simulator");
-#else
                 // Exclude libraries that don't match the target SDK
                 if (PlayerSettings.VisionOS.sdkVersion == VisionOSSdkVersion.Device)
                 {
@@ -95,8 +107,8 @@ namespace UnityEditor.XR.VisionOS
                 }
 
                 return true;
-#endif
             }
         }
+#endif
     }
 }