+++ Runtime/Plugins/visionos/UnityVisionOS.m
@@ -1,5 +1,6 @@
 #import <Foundation/Foundation.h>
 #include "IUnityInterface.h"
+#include "UnityVisionOSParameters.h"
 #import <CompositorServices/CompositorServices.h>
 
 #define EXPORT(RETURN_TYPE) RETURN_TYPE __attribute__ ((visibility("default")))  __attribute__((__used__))
@@ -12,6 +13,8 @@ void UNITY_INTERFACE_EXPORT UNITY_INTERFACE_API UnityPluginLoad(IUnityInterfaces
 
 void ObjC_SetCPLayer(cp_layer_renderer_t layer);
 
+void ObjC_SetDisplayProviderParameters(void* parameters);
+
 #ifdef __cplusplus
 } // extern "C"
 #endif
@@ -21,6 +24,8 @@ void ObjC_SetCPLayer(cp_layer_renderer_t layer);
 + (void)loadPlugin;
 
 + (void)setLayerRenderer:(cp_layer_renderer_t)layerRenderer;
+
++ (void)setDisplayProviderParameters:(NSValue*)parameters;
 
 + (NSNumber*)getSinglePass;
 
@@ -36,6 +41,13 @@ void ObjC_SetCPLayer(cp_layer_renderer_t layer);
 + (void)setLayerRenderer:(cp_layer_renderer_t)layerRenderer
 {
     ObjC_SetCPLayer(layerRenderer);
+}
+
++ (void)setDisplayProviderParameters:(NSValue*)parameters
+{
+    DisplayProviderParameters params;
+    [parameters getValue:&params size:sizeof(DisplayProviderParameters)];
+    ObjC_SetDisplayProviderParameters(&params);
 }
 
 + (NSNumber*)getSinglePass